requirements.txt
flask==3.0.0
flask-login==0.6.3
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
werkzeug==3.0.1
bson==0.5.10
pandas==2.1.4
plotly==5.18.0
dash==2.14.2
numpy==1.26.2
scikit-learn==1.3.2 